esp32s3-usbserial: common usbserial for both esp32c3 & esp32s3
